Add better input handling for Fram MPI
If invalid input is given for minutes, hours or days, the program
repeats only the single wrong input instead of asking for all numbers
again. Interval check is also implemented to specify a valid number of
minutes, hours and days.

'n_nodes' is removed from the global scope, and the global declaration
is removed from main(). The value is not used globally and should
therefore not be defined globally.