Step 1 of improve log file reading
transit.F90 is modified to print with higher decimal precision. This program used substring indexing to fetch values so it had to be updated. When this process is done, only splits are used to make the process more flexible and to be compatible with future potential changes to the decimal precisions.